20170227 - updating a pastor with no assigned pastor was creating a relationship record with a pastor id of 0 that was causing trying to get property of non-object errors, these relationships need to be deleted

+++ app/Http/Controllers/ParishesController.php
@@ -314,9 +314,12 @@ return Redirect::action('ParishesController@index');
             $pastor->contact_id_b = $request->input('pastor_id');
             $pastor->relationship_type_id = RELATIONSHIP_TYPE_PASTOR;
             $pastor->is_active = 1;
-            $pastor->save();
+            // avoid creating relationship if no pastor is assigned
+            if ($pastor->contact_id_b > 0) {
+                $pastor->save();
+            }
             // if there is no pastor assigned then delete the relationship
-            if ($pastor->contact_id_b == 0) {
+            if (($pastor->contact_id_b == 0) && (isset($pastor->id))) {
                 $pastor->delete();
             }
